Add a UNKNOWN_ERROR variant first.
The prost library we're using chooses the first
defined variant when converting from an i32 that has
an invalid value. This would happen when a client
running an older version of swbus receives a newer
error code. We want that to be recognized as an error,
not converted implicitly to SWBUS_ERROR_CODE_OK.