Hook up to swss-common logger (#45)
### Why
SONiC uses config_db to keep logger configuration, which allows changing
the configuration on the fly. On the back, swss-common logger subscribes
to config_db logger table and notifies the corresponding component.

### What this PR does
Hook up to swss-common logger to receive configuration update. In
response to the change, dynamically changing log level in rust.

This PR can only be committed after
https://github.com/sonic-net/sonic-swss-common/pull/969