Add rust implementation of container Python script (#38)
Add a rust implementation of the `container` script in the SONiC image.
This script is primarily responsible for starting, waiting on, and
stopping containers.

In the case of local containers, this script is basically a wrapper
around `docker start`, `docker wait`, and `docker stop`. In the case of
Kubernetes containers, there's some additional logic/checks on whether
the container should be started or not; this is not fully implemented.

Basic testing has been done with local containers (the most common use
case).

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>