Make deserializer more forgiven to parse dash_bfd_probe_state (#113)
### why
currently the deserializer for dash_bfd_probe_state has strict
requirements on the format of the fields. If it doesn't follow the
format, it will reject it. Specifically, the timestamp field is enclosed
in double-quotes, which caused parsing error.

### what this PR does
make the deserializer more forgiven with format. If the value has double
or single quotes or whitespaces, remove them first. If the value of
v4_bfd_up_sessions or v6_bfd_up_sessions has quotes or space between
comma, remove them.