Fix issue #118: use hostname to build service path (#122)
### why
swbus-cli show swbusd route is not working. The issue is introduced by
PR https://github.com/sonic-net/sonic-dash-ha/pull/108. In the PR, it
compares the dest sp of received ManagementRequest to "my_service_path".
the request is processed if they are same. In a setup with both ipv4 and
ipv6 endpoints, my_service_path takes one of them. If the dest sp of the
request picks the other one, they don't match so the request will be
dropped.

### what this PR does
1. we use hostname in DEVICE_META|localhost to build my service path. 
2. remote service path is exchanged when connection is established.
Client sends its own service path in the request and server replies with
its one when connection is created.